Implementation du module Reglement:Modification d'un paiement

+++ src/main/java/com/pixel/synchronre/sychronremodule/model/dto/reglement/response/PaiementDetailsResp.java
@@ -2,6 +2,7 @@ package com.pixel.synchronre.sychronremodule.model.dto.reglement.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
@@ -9,8 +10,8 @@ public class PaiementDetailsResp {
     private Long regId;
     private String regReference;
     private LocalDate regDate;
-    private float regMontant;
-    private float regCommission;
+    private BigDecimal regMontant;
+    private BigDecimal regCommission;
     private String typeReglement;
     private String affCode;
     private String affAssure;
